feat: 관리자 API 엔드포인트 추가
