+++ Projects/Data/Sources/Remote/PopupAPI.swift
@@ -1,124 +1,6 @@
 import Core
-import Domain
 import Foundation
 import Moya
-
-public enum AdminAPI {
-    case deactivatePopup(userUuid: String, popupUuid: String)
-}
-
-extension AdminAPI: BaseAPI {
-    public var path: String {
-        switch self {
-        case .deactivatePopup(let userUuid, let popupUuid):
-            return "/admin/user/\(userUuid)/popup/\(popupUuid)/deactivate"
-        }
-    }
-
-    public var method: Moya.Method {
-        switch self {
-        case .deactivatePopup:
-            return .patch
-        }
-    }
-
-    public var task: Task {
-        .requestPlain
-    }
-}
-
-public enum AppleAuthAPI {
-    case loginWithEmail(authCode: String, email: String)
-    case login(authCode: String)
-    case signup(userDto: UserDTO)
-}
-
-extension AppleAuthAPI: BaseAPI {
-    public var path: String {
-        switch self {
-        case .loginWithEmail, .login:
-            return "/auth/apple/mobile/login"
-        case .signup:
-            return "/auth/apple/signup"
-        }
-    }
-
-    public var method: Moya.Method {
-        .post
-    }
-
-    public var task: Task {
-        switch self {
-        case .loginWithEmail(let authCode, let email):
-            return .requestJSONEncodable([
-                "auth_code": authCode,
-                "email": email,
-            ])
-        case .login(let authCode):
-            return .requestJSONEncodable(["auth_code": authCode])
-        case .signup(let userDto):
-            return .requestJSONEncodable(userDto)
-        }
-    }
-}
-
-public enum GoogleAuthAPI {
-    case login(idToken: String)
-    case signup(userDTO: UserDTO)
-}
-
-extension GoogleAuthAPI: BaseAPI {
-    public var path: String {
-        switch self {
-        case .login:
-            return "/auth/google/mobile/login"
-        case .signup:
-            return "/auth/google/signup"
-        }
-    }
-
-    public var method: Moya.Method {
-        .post
-    }
-
-    public var task: Task {
-        switch self {
-        case .login(let idToken):
-            return .requestJSONEncodable(["id_token": idToken])
-        case .signup(let userDTO):
-            return .requestJSONEncodable(userDTO)
-        }
-    }
-}
-
-public enum KakaoAuthAPI {
-    case login(accessToken: String)
-    case signup(userDTO: UserDTO)
-}
-
-extension KakaoAuthAPI: BaseAPI {
-    public var path: String {
-        switch self {
-        case .login:
-            return "/auth/kakao/mobile/login"
-        case .signup:
-            return "/auth/kakao/signup"
-        }
-    }
-
-    public var method: Moya.Method {
-        .post
-    }
-
-    public var task: Task {
-        switch self {
-        case .login(let accessToken):
-            return .requestJSONEncodable(["access_token": accessToken])
-        case .signup(let userDTO):
-            return .requestJSONEncodable(userDTO)
-        }
-    }
-}
 
 public enum PopupAPI {
     case getPopupList
@@ -292,97 +174,6 @@ extension PopupAPI: BaseAPI {
                 ],
                 encoding: JSONEncoding.default
             )
-        }
-    }
-}
-
-public enum UserAPI {
-    case checkNickname(nickname: String)
-    case updateNickname(userUuid: String, newNickname: String)
-    case autoLogin(userUuid: String)
-    case getRecommendList
-    case hardDeleteUser(userUuid: String)
-    case getAlertKeywordList(userUuid: String)
-    case addAlertKeyword(userUuid: String, newAlertKeyword: String)
-    case removeAlertKeyword(userUuid: String, deleteAlertKeyword: String)
-    case alertStatus(userUuid: String, isAlerted: Bool)
-    case checkFcmToken(userUuid: String, fcmToken: String)
-    case updateFcmToken(userUuid: String, newFcmToken: String)
-}
-
-extension UserAPI: BaseAPI {
-    public var path: String {
-        switch self {
-        case .checkNickname:
-            return "/user/nickname/duplicated"
-        case .updateNickname(let userUuid, _):
-            return "/user/\(userUuid)"
-        case .autoLogin:
-            return "/auth/autoLogin"
-        case .getRecommendList:
-            return "/recommend"
-        case .hardDeleteUser(let userUuid):
-            return "/user/\(userUuid)/hard-delete"
-        case .getAlertKeywordList, .addAlertKeyword, .removeAlertKeyword:
-            return "/alert-keyword"
-        case .alertStatus(let userUuid, _):
-            return "/user/\(userUuid)/alert-status"
-        case .checkFcmToken(let userUuid, _):
-            return "/user/\(userUuid)/fcm-token/duplicate-check"
-        case .updateFcmToken(let userUuid, _):
-            return "/user/\(userUuid)/fcm-token/update"
-        }
-    }
-
-    public var method: Moya.Method {
-        switch self {
-        case .checkNickname, .getRecommendList, .getAlertKeywordList, .checkFcmToken:
-            return .get
-        case .autoLogin, .addAlertKeyword:
-            return .post
-        case .updateNickname, .alertStatus:
-            return .patch
-        case .updateFcmToken:
-            return .put
-        case .hardDeleteUser, .removeAlertKeyword:
-            return .delete
-        }
-    }
-
-    public var task: Task {
-        switch self {
-        case .checkNickname(let nickname):
-            return .requestParameters(parameters: ["nickname": nickname], encoding: URLEncoding.queryString)
-        case .updateNickname(_, let newNickname):
-            return .requestParameters(parameters: ["nickname": newNickname], encoding: JSONEncoding.default)
-        case .autoLogin(let userUuid):
-            return .requestJSONEncodable(["userUuid": userUuid])
-        case .getRecommendList, .hardDeleteUser:
-            return .requestPlain
-        case .getAlertKeywordList(let userUuid):
-            return .requestParameters(parameters: ["userUuid": userUuid], encoding: URLEncoding.queryString)
-        case .addAlertKeyword(let userUuid, let newAlertKeyword):
-            return .requestParameters(
-                parameters: [
-                    "userUuid": userUuid,
-                    "newAlertKeyword": newAlertKeyword,
-                ],
-                encoding: JSONEncoding.default
-            )
-        case .removeAlertKeyword(let userUuid, let deleteAlertKeyword):
-            return .requestParameters(
-                parameters: [
-                    "userUuid": userUuid,
-                    "deleteAlertKeyword": deleteAlertKeyword,
-                ],
-                encoding: JSONEncoding.default
-            )
-        case .alertStatus(_, let isAlerted):
-            return .requestParameters(parameters: ["isAlerted": isAlerted], encoding: JSONEncoding.default)
-        case .checkFcmToken(_, let fcmToken):
-            return .requestParameters(parameters: ["fcmToken": fcmToken], encoding: URLEncoding.queryString)
-        case .updateFcmToken(_, let newFcmToken):
-            return .requestParameters(parameters: ["fcmToken": newFcmToken], encoding: JSONEncoding.default)
         }
     }
 }
